modbuilder + workshop
Modified the ModBuilder to allow building with Newtonsoft.Json
Modified the ModBuilder to allow creating the lib folder on first launch
Added Workshop GUI, currently on the old MLL api (the levels don't work obv)
Moved some calls to the api to threaded code, so game doesn't hang waiting for requests to complete.
Added a "Loading x levels.." so users know why the workshop takes so long to load.
+ Other small changes